+++ src/exporter/jsonl.rs
@@ -92,7 +92,7 @@ impl JsonlExporter {
             statement: meta.statement.to_string(),
             appname: meta.appname.to_string(),
             client_ip: meta.client_ip.to_string(),
-            tag: sqllog.tag.as_ref().map(|t| t.to_string()),
+            tag: sqllog.tag.as_ref().map(ToString::to_string),
             sql: sqllog.body().to_string(),
             exec_time_ms: ind.as_ref().map(|i| i.execute_time),
             row_count: ind.as_ref().map(|i| i.row_count),