+++ tests/filter_meta_tests.rs
@@ -51,8 +51,6 @@ directory = "{}"
 
 [features.filters]
 enable = true
-
-[features.filters.meta]
 sess_ids = ["SESSION1"]
 
 [exporter.csv]
@@ -97,8 +95,6 @@ directory = "{}"
 
 [features.filters]
 enable = true
-
-[features.filters.meta]
 thrd_ids = ["THREAD2"]
 
 [exporter.csv]
@@ -143,8 +139,6 @@ directory = "{}"
 
 [features.filters]
 enable = true
-
-[features.filters.meta]
 usernames = ["USER3"]
 
 [exporter.csv]
@@ -189,8 +183,6 @@ directory = "{}"
 
 [features.filters]
 enable = true
-
-[features.filters.meta]
 statements = ["UPD", "SEL"]
 
 [exporter.csv]
@@ -236,8 +228,6 @@ directory = "{}"
 
 [features.filters]
 enable = true
-
-[features.filters.meta]
 appnames = ["APP1"]
 
 [exporter.csv]